Split the check catalog and results model files
