--- web/models/check.go
@@ -1,109 +1,0 @@
-package models
-
-import (
-	"sort"
-	"time"
-
-	"gorm.io/datatypes"
-)
-
-const (
-	CheckPassing   string = "passing"
-	CheckWarning   string = "warning"
-	CheckCritical  string = "critical"
-	CheckSkipped   string = "skipped"
-	CheckUndefined string = "undefined"
-)
-
-// List is used instead of a map as it guarantees order
-type ChecksCatalog []*Check
-
-type Check struct {
-	ID             string `json:"id,omitempty"`
-	Name           string `json:"name,omitempty"`
-	Group          string `json:"group,omitempty"`
-	Description    string `json:"description,omitempty"`
-	Remediation    string `json:"remediation,omitempty"`
-	Implementation string `json:"implementation,omitempty"`
-	Labels         string `json:"labels,omitempty"`
-	Selected       bool   `json:"selected,omitempty"`
-	Result         string `json:"result,omitempty"`
-}
-
-// Store the data as payload. Changes in this struct are expected
-type CheckRaw struct {
-	ID        string `gorm:"primaryKey"`
-	CreatedAt time.Time
-	Payload   datatypes.JSON
-}
-
-// TableName overrides the table name used by CheckRaw to `checks`
-func (CheckRaw) TableName() string {
-	return "checks"
-}
-
-type GroupedChecks struct {
-	Group  string
-	Checks ChecksCatalog
-}
-
-type GroupedCheckList []*GroupedChecks
-
-type CheckResultsRaw struct {
-	ID        int64
-	CreatedAt time.Time
-	GroupID   string
-	Payload   datatypes.JSON
-}
-
-// TableName overrides the table name used by CheckResultsRaw to `check_results`
-func (CheckResultsRaw) TableName() string {
-	return "check_results"
-}
-
-type Results struct {
-	Hosts  map[string]*Host         `json:"hosts,omitempty"`
-	Checks map[string]*ChecksByHost `json:"checks,omitempty"`
-}
-
-// The ChecksByHost struct stores the checks list, but the results are grouped by hosts
-type ChecksByHost struct {
-	Hosts map[string]*Check `json:"hosts,omitempty"`
-}
-
-type Host struct {
-	Reachable bool   `json:"reachable"`
-	Msg       string `json:"msg"`
-}
-
-// Simplified models for the frontend
-type ClusterCheckResults struct {
-	Hosts  map[string]*Host     `json:"hosts"`
-	Checks []ClusterCheckResult `json:"checks"`
-}
-
-type ClusterCheckResult struct {
-	ID          string            `json:"id,omitempty"`
-	Hosts       map[string]*Check `json:"hosts,omitempty"`
-	Group       string            `json:"group,omitempty"`
-	Description string            `json:"description,omitempty"`
-}
-
-// Sorting methods for GroupedCheckList
-
-func (g GroupedCheckList) Len() int {
-	return len(g)
-}
-
-func (g GroupedCheckList) Less(i, j int) bool {
-	return g[i].Checks[0].Name < g[j].Checks[0].Name
-}
-
-func (g GroupedCheckList) Swap(i, j int) {
-	g[i], g[j] = g[j], g[i]
-}
-
-func (g GroupedCheckList) OrderByName() GroupedCheckList {
-	sort.Sort(g)
-	return g
-}
